+++ csvtk/cmd/uniq.go
@@ -1,0 +1,135 @@
+// Copyright © 2016-2023 Wei Shen <shenwei356@gmail.com>
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package cmd
+
+import (
+	"encoding/csv"
+	"fmt"
+	"runtime"
+	"strings"
+
+	"github.com/shenwei356/xopen"
+	"github.com/spf13/cobra"
+)
+
+// uniqCmd represents the uniq command
+var uniqCmd = &cobra.Command{
+	Use:   "uniq",
+	Short: "unique data without sorting",
+	Long: `unique data without sorting
+
+`,
+	Run: func(cmd *cobra.Command, args []string) {
+		config := getConfigs(cmd)
+		files := getFileListFromArgsAndFile(cmd, args, true, "infile-list", true)
+		if len(files) > 1 {
+			checkError(fmt.Errorf("no more than one file should be given"))
+		}
+
+		runtime.GOMAXPROCS(config.NumCPUs)
+
+		fieldStr := getFlagString(cmd, "fields")
+		if fieldStr == "" {
+			checkError(fmt.Errorf("flag -f (--fields) needed"))
+		}
+
+		fuzzyFields := getFlagBool(cmd, "fuzzy-fields")
+		ignoreCase := getFlagBool(cmd, "ignore-case")
+		keepN := getFlagPositiveInt(cmd, "keep-n")
+
+		outfh, err := xopen.Wopen(config.OutFile)
+		checkError(err)
+		defer outfh.Close()
+
+		writer := csv.NewWriter(outfh)
+		if config.OutTabs || config.Tabs {
+			if config.OutDelimiter == ',' {
+				writer.Comma = '\t'
+			} else {
+				writer.Comma = config.OutDelimiter
+			}
+		} else {
+			writer.Comma = config.OutDelimiter
+		}
+
+		keysMaps := make(map[string]int, 10000)
+
+		file := files[0]
+		csvReader, err := newCSVReaderByConfig(config, file)
+
+		if err != nil {
+			if err == xopen.ErrNoContent {
+				log.Warningf("csvtk uniq: skipping empty input file: %s", file)
+
+				writer.Flush()
+				checkError(writer.Error())
+				readerReport(&config, csvReader, file)
+				return
+			}
+			checkError(err)
+		}
+
+		csvReader.Read(ReadOption{
+			FieldStr:    fieldStr,
+			FuzzyFields: fuzzyFields,
+
+			DoNotAllowDuplicatedColumnName: true,
+		})
+
+		var key string
+		var n int
+		var ok bool
+
+		for record := range csvReader.Ch {
+			if record.Err != nil {
+				checkError(record.Err)
+			}
+
+			key = strings.Join(record.Selected, "_shenwei356_")
+			if ignoreCase {
+				key = strings.ToLower(key)
+			}
+			if n, ok = keysMaps[key]; ok {
+				if n >= keepN {
+					continue
+				}
+				keysMaps[key]++
+			} else {
+				keysMaps[key] = 1
+			}
+			checkError(writer.Write(record.All))
+		}
+
+		writer.Flush()
+		checkError(writer.Error())
+
+		readerReport(&config, csvReader, file)
+	},
+}
+
+func init() {
+	RootCmd.AddCommand(uniqCmd)
+	uniqCmd.Flags().StringP("fields", "f", "1", `select these fields as keys. e.g -f 1,2 or -f columnA,columnB`)
+	uniqCmd.Flags().BoolP("ignore-case", "i", false, `ignore case`)
+	uniqCmd.Flags().BoolP("fuzzy-fields", "F", false, `using fuzzy fields, e.g., -F -f "*name" or -F -f "id123*"`)
+	uniqCmd.Flags().IntP("keep-n", "n", 1, `keep at most N records for a key`)
+
+}