+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 - [csvtk v0.27.0](https://github.com/shenwei356/csvtk/releases/tag/v0.27.0)
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/csvtk/v0.27.0/total.svg)](https://github.com/shenwei356/csvtk/releases/tag/v0.27.0)
   - add a new global flag `-Z/--show-row-number`, supported commands: cut, pretty, csv2tab, csv2xlsx, tab2csv
-  - `csvtk csv2json/filter2/fold/freq/gater/grep/inter/join/mutate/mutate2/pretty/sep/sort/split/splitxlsx/summary/unfold/uniq`:
+  - `csvtk csv2json/uniq`:
       - only selected column names are not allowed to be duplicated in the input data. [#235](https://github.com/shenwei356/csvtk/issues/235)
   - `csvtk mutate/mutate2`:
       - **new flags `--at`, `--before`, `--after` for specifying the position of the new column**. [#193](https://github.com/shenwei356/csvtk/issues/193)